Refactor types generation
* Follow DirectedCodegen's "golden-path" during shape generation
* Remove the use indirection and visitor in generators
* Created specific generators handling each shape directive
* Configure DirectedCodegen to "feed" shapes in alphabetical ordering based on shape name
* Upgrade Smithy version
