+++ codegen/projections/high_score_service/lib/high_score_service/types.rb
@@ -9,6 +9,7 @@
 
 module HighScoreService
   module Types
+
     # Input structure for CreateHighScore
     #
     # @!attribute high_score
