+++ codegen/projections/weather/lib/weather/types.rb
@@ -584,6 +584,7 @@ module Weather
 
       ULTRA = 3
     end
+
     # Includes enum constants for TypedYesNo
     #
     module TypedYesNo
