+++ codegen/smithy-ruby-codegen/src/main/java/software/amazon/smithy/ruby/codegen/DirectedRubyCodegen.java
@@ -15,7 +15,6 @@
 
 package software.amazon.smithy.ruby.codegen;
 
-import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Comparator;
 import java.util.HashSet;
@@ -38,17 +37,19 @@ import software.amazon.smithy.codegen.core.directed.GenerateStructureDirective;
 import software.amazon.smithy.codegen.core.directed.GenerateUnionDirective;
 import software.amazon.smithy.model.Model;
 import software.amazon.smithy.model.shapes.ServiceShape;
-import software.amazon.smithy.model.shapes.Shape;
 import software.amazon.smithy.model.shapes.ShapeId;
 import software.amazon.smithy.ruby.codegen.config.ClientConfig;
 import software.amazon.smithy.ruby.codegen.generators.ClientGenerator;
 import software.amazon.smithy.ruby.codegen.generators.ConfigGenerator;
+import software.amazon.smithy.ruby.codegen.generators.EnumGenerator;
 import software.amazon.smithy.ruby.codegen.generators.GemspecGenerator;
 import software.amazon.smithy.ruby.codegen.generators.HttpProtocolTestGenerator;
+import software.amazon.smithy.ruby.codegen.generators.IntEnumGenerator;
 import software.amazon.smithy.ruby.codegen.generators.ModuleGenerator;
 import software.amazon.smithy.ruby.codegen.generators.PaginatorsGenerator;
 import software.amazon.smithy.ruby.codegen.generators.ParamsGenerator;
-import software.amazon.smithy.ruby.codegen.generators.TypesGenerator;
+import software.amazon.smithy.ruby.codegen.generators.StructureGenerator;
+import software.amazon.smithy.ruby.codegen.generators.UnionGenerator;
 import software.amazon.smithy.ruby.codegen.generators.ValidatorsGenerator;
 import software.amazon.smithy.ruby.codegen.generators.WaitersGenerator;
 import software.amazon.smithy.ruby.codegen.generators.YardOptsGenerator;
@@ -60,7 +61,10 @@ public class DirectedRubyCodegen
     private static final Logger LOGGER =
             Logger.getLogger(DirectedRubyCodegen.class.getName());
 
-    private List<Shape> typeShapes = new ArrayList<>();
+    private final StructureGenerator structureGenerator = new StructureGenerator();
+    private final UnionGenerator unionGenerator = new UnionGenerator();
+    private final EnumGenerator enumGenerator = new EnumGenerator();
+    private final IntEnumGenerator intEnumGenerator = new IntEnumGenerator();
 
     @Override
     public SymbolProvider createSymbolProvider(CreateSymbolProviderDirective<RubySettings> directive) {
@@ -152,8 +156,27 @@ public class DirectedRubyCodegen
     }
 
     @Override
+    public void customizeBeforeShapeGeneration(CustomizeDirective<GenerationContext, RubySettings> directive) {
+        // Pre-populate module blocks to types.rb and types.rbs files
+        var settings = directive.context().settings();
+        var namespace = settings.getModule() + "::Types";
+        var rbFile = settings.getGemName() + "/lib/" + settings.getGemName() + "/types.rb";
+        var rbsFile = settings.getGemName() + "/sig/" + settings.getGemName() + "/types.rbs";
+        directive.context().writerDelegator().useFileWriter(rbFile, namespace, writer -> {
+            writer.includePreamble().includeRequires();
+            writer.addModule(settings.getModule());
+            writer.addModule("Types");
+        });
+        directive.context().writerDelegator().useFileWriter(rbsFile, namespace, writer -> {
+            writer.includePreamble();
+            writer.addModule(settings.getModule());
+            writer.addModule("Types");
+        });
+    }
+
+    @Override
     public void generateStructure(GenerateStructureDirective<GenerationContext, RubySettings> directive) {
-        typeShapes.add(directive.shape());
+        structureGenerator.accept(directive);
     }
 
     @Override
@@ -161,54 +184,29 @@ public class DirectedRubyCodegen
         if (directive.context().protocolGenerator().isPresent()) {
             directive.context().protocolGenerator().get().generateErrors(directive.context());
         }
-        typeShapes.add(directive.shape());
+        structureGenerator.accept(directive);
     }
 
     @Override
     public void generateUnion(GenerateUnionDirective<GenerationContext, RubySettings> directive) {
-        typeShapes.add(directive.shape());
+        unionGenerator.accept(directive);
     }
 
     @Override
     public void generateEnumShape(GenerateEnumDirective<GenerationContext, RubySettings> directive) {
-        typeShapes.add(directive.shape());
+        enumGenerator.accept(directive);
     }
 
     @Override
     public void generateIntEnumShape(GenerateIntEnumDirective<GenerationContext, RubySettings> directive) {
-        typeShapes.add(directive.shape());
+        intEnumGenerator.accept(directive);
     }
 
     @Override
     public void customizeBeforeIntegrations(CustomizeDirective<GenerationContext, RubySettings> directive) {
         GenerationContext context = directive.context();
-
-        // Generate types
-        TypesGenerator typesGenerator = new TypesGenerator(context);
-        context.writerDelegator().useFileWriter(
-            typesGenerator.getFile(), typesGenerator.getNameSpace(), writer -> {
-            writer.includePreamble().includeRequires();
-
-            TypesGenerator.TypesVisitor visitor = typesGenerator.getTypeVisitor(writer);
-
-            writer.addModule(directive.settings().getModule());
-            writer.addModule("Types");
-
-            typeShapes.stream()
-                .sorted(Comparator.comparing(a -> a.getId().getName()))
-                .forEach(shape -> shape.accept(visitor));
-
-            writer.closeAllModules();
-        });
-
-        typesGenerator.renderRbs();
-
-        ParamsGenerator paramsGenerator = new ParamsGenerator(context);
-        paramsGenerator.render();
-
-        ValidatorsGenerator validatorsGenerator = new ValidatorsGenerator(context);
-        validatorsGenerator.render();
-
+        new ParamsGenerator(context).render();
+        new ValidatorsGenerator(context).render();
 
         if (directive.context().protocolGenerator().isPresent()) {
             ProtocolGenerator generator = directive.context().protocolGenerator().get();
@@ -239,6 +237,17 @@ public class DirectedRubyCodegen
                     new HttpProtocolTestGenerator(context);
             testGenerator.render();
         }
+    }
+
+    @Override
+    public void customizeAfterIntegrations(CustomizeDirective<GenerationContext, RubySettings> directive) {
+        // Close all module blocks to types.rb and types.rbs files
+        var settings = directive.context().settings();
+        var namespace = settings.getModule() + "::Types";
+        var rbFile = settings.getGemName() + "/lib/" + settings.getGemName() + "/types.rb";
+        var rbsFile = settings.getGemName() + "/sig/" + settings.getGemName() + "/types.rbs";
+        directive.context().writerDelegator().useFileWriter(rbFile, namespace, RubyCodeWriter::closeAllModules);
+        directive.context().writerDelegator().useFileWriter(rbsFile, namespace, RubyCodeWriter::closeAllModules);
     }
 
     private Set<RubyDependency> collectDependencies(
