+++ codegen/smithy-ruby-codegen/src/main/java/software/amazon/smithy/ruby/codegen/generators/UnionGenerator.java
@@ -1,0 +1,121 @@
+/*
+ * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License").
+ * You may not use this file except in compliance with the License.
+ * A copy of the License is located at
+ *
+ *  http://aws.amazon.com/apache2.0
+ *
+ * or in the "license" file accompanying this file. This file is distributed
+ * on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
+ * express or implied. See the License for the specific language governing
+ * permissions and limitations under the License.
+ */
+
+package software.amazon.smithy.ruby.codegen.generators;
+
+import java.util.function.Consumer;
+import software.amazon.smithy.codegen.core.Symbol;
+import software.amazon.smithy.codegen.core.SymbolProvider;
+import software.amazon.smithy.codegen.core.directed.GenerateUnionDirective;
+import software.amazon.smithy.model.Model;
+import software.amazon.smithy.model.shapes.MemberShape;
+import software.amazon.smithy.model.traits.SensitiveTrait;
+import software.amazon.smithy.ruby.codegen.GenerationContext;
+import software.amazon.smithy.ruby.codegen.Hearth;
+import software.amazon.smithy.ruby.codegen.RubyCodeWriter;
+import software.amazon.smithy.ruby.codegen.RubyFormatter;
+import software.amazon.smithy.ruby.codegen.RubySettings;
+import software.amazon.smithy.ruby.codegen.generators.docs.ShapeDocumentationGenerator;
+
+public final class UnionGenerator
+        implements Consumer<GenerateUnionDirective<GenerationContext, RubySettings>> {
+    @Override
+    public void accept(GenerateUnionDirective<GenerationContext, RubySettings> directive) {
+        var model = directive.model();
+        var settings = directive.context().settings();
+        var namespace = settings.getModule() + "::Types";
+        var rbFile = settings.getGemName() + "/lib/" + settings.getGemName() + "/types.rb";
+        var rbsFile = settings.getGemName() + "/sig/" + settings.getGemName() + "/types.rbs";
+        var shape = directive.shape();
+        var symbolProvider = directive.context().symbolProvider();
+        String documentation = new ShapeDocumentationGenerator(model, symbolProvider, shape).render();
+
+        directive.context().writerDelegator().useFileWriter(rbFile, namespace, writer -> {
+            writer.writeInline("$L", documentation);
+            writer.openBlock("class $T < $T", symbolProvider.toSymbol(shape), Hearth.UNION);
+
+            for (MemberShape memberShape : shape.members()) {
+                String memberDocumentation =
+                        new ShapeDocumentationGenerator(model, symbolProvider, memberShape).render();
+
+                writer
+                        .writeInline("$L", memberDocumentation)
+                        .openBlock("class $L < $T",
+                                symbolProvider.toMemberName(memberShape), symbolProvider.toSymbol(shape))
+                        .openBlock("def to_h")
+                        .write("{ $L: super(__getobj__) }",
+                                RubyFormatter.toSnakeCase(symbolProvider.toMemberName(memberShape)))
+                        .closeBlock("end")
+                        .call(() -> renderUnionToSMethod(settings, symbolProvider, writer, model, memberShape))
+                        .closeBlock("end\n");
+            }
+
+            writer
+                    .writeDocstring("Handles unknown future members")
+                    .openBlock("class Unknown < $T", symbolProvider.toSymbol(shape))
+                    .openBlock("def to_h")
+                    .write("{ unknown: super(__getobj__) }")
+                    .closeBlock("end\n")
+                    .openBlock("def to_s")
+                    .write("\"#<$L::Types::Unknown #{__getobj__ || 'nil'}>\"", settings.getModule())
+                    .closeBlock("end")
+                    .closeBlock("end")
+                    .closeBlock("end\n");
+        });
+
+        directive.context().writerDelegator().useFileWriter(rbsFile, namespace, writer -> {
+            Symbol symbol = symbolProvider.toSymbol(shape);
+            writer.openBlock("class $T < $T", symbol, Hearth.UNION);
+
+            for (MemberShape memberShape : shape.members()) {
+                writer
+                    .openBlock("class $L < $T",
+                            symbolProvider.toMemberName(memberShape), symbol)
+                    .write("def to_h: () -> { $L: Hash[Symbol,$T] }",
+                            RubyFormatter.toSnakeCase(symbolProvider.toMemberName(memberShape)),
+                            symbol)
+                    .closeBlock("end\n");
+            }
+
+            writer
+                .openBlock("class Unknown < $T", symbol)
+                .write("def to_h: () -> { unknown: Hash[Symbol,$T] }",
+                        symbol)
+                .closeBlock("end")
+                .closeBlock("end\n");
+        });
+    }
+
+    private void renderUnionToSMethod(
+        RubySettings settings,
+        SymbolProvider symbolProvider,
+        RubyCodeWriter writer,
+        Model model,
+        MemberShape memberShape) {
+        String fullQualifiedShapeName = settings.getModule() + "::Types::"
+                + symbolProvider.toMemberName(memberShape);
+
+        writer.write("")
+                .openBlock("def to_s");
+
+        if (memberShape.getMemberTrait(model, SensitiveTrait.class).isPresent()) {
+            writer.write("\"#<$L [SENSITIVE]>\"", fullQualifiedShapeName);
+        } else {
+            writer.write("\"#<$L #{__getobj__ || 'nil'}>\"", fullQualifiedShapeName);
+        }
+
+        writer.closeBlock("end");
+    }
+}
